pyproject.toml
[build-system]
requires = [
  "setuptools>=64",
  "setuptools-scm>=8",
  "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fusion_surrogates"
description = "Transport surrogate models for Tokamak fusion."
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
license-files = ["LICENSE"]
authors = [{name = "DeepMind"}]
classifiers = [
    "Intended Audience :: Science/Research",
]

dependencies = [
    "absl-py>=2.0.0",
    "jax>=0.4.31",
    "flax>=0.7.5",
    "numpy>2",
    "setuptools;python_version>='3.10'",
    "immutabledict",
]

[project.optional-dependencies]
testing = [
    "pytest",
    "pytest-xdist",
    "pylint>=2.6.0",
    "pyink",
]
tglfnnukaea = [
    "torch",
    "pyyaml",
]

[tool.setuptools.packages.find]
include = ["fusion_surrogates*"]
